+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to the "go-outliner" extension will be documented in this file.
 
+## 0.1.6 - 2018-05-08
+### Changed
+- Moved UI to separate tab
+
+### Added
+- Tests/Benchmarks view added
+    - Test/Benchmark specific function command added
+    - Test/Benchmark all functions command added
+
 ## 0.0.6 - 2018-05-08
 ### Changed
 - Items are now sorted by label